+++ libraries/SuplaDevice/src/supla/tools.cpp
