+++ libraries/SuplaDevice/src/supla/tools.h
